+++ backend/foodtruck/src/main/java/com/connectruck/foodtruck/notification/infra/OrderMessageHandler.java
@@ -1,0 +1,8 @@
+package com.connectruck.foodtruck.notification.infra;
+
+import com.connectruck.foodtruck.order.message.OrderMessage;
+
+public interface OrderMessageHandler {
+
+    void handle(OrderMessage orderMessage);
+}
